third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-valid-relative-color-expected.txt

This is a testharness.js-based test.
Found 71 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['color'] = "oklab(from color-mix(in oklab, oklab(0.25 0.2 0.5), oklab(0.25 0.2 0.5)) l a b / alpha)" should set the property value
  Colors do not match.\nActual:   oklab(from color-mix(in oklab, oklab(0.25 0.2 0.5), oklab(0.25 0.2 0.5)) l a b / alpha)\nExpected: oklab(from color-mix(oklab(0.25 0.2 0.5), oklab(0.25 0.2 0.5)) l a b / alpha).\nError: assert_equals: Color format is correct. expected "oklab(from color-mix(oklab(  ), oklab(  )) l a b / alpha)" but got "oklab(from color-mix(in oklab, oklab(  ), oklab(  )) l a b / alpha)"
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 0 0)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 0 0 / 0)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 0 b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 0 / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 0)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 0 g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 0 b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 0 / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 0)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0.2 g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 20% g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 0.2 b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 20% b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 0.2 / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 20% / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 0.2)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 20%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 0.2 g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 20% g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 0.2 b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 20% b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 0.2 / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 20% / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 0.2)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 20%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 2 3 4)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 2 3 4 / 5)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -2 -3 -4)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -2 -3 -4 / -5)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 200% 300% 400%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 200% 300% 400% / 500%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -200% -300% -400%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -200% -300% -400% / -500%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear g b r)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear b alpha r / g)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r r r / r)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear alpha alpha alpha / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear g b r)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear b alpha r / g)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r r r / r)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear alpha alpha alpha / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 1.7 1.5 1.3) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 1.7 1.5 1.3) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 1.7 1.5 1.3 / 140%) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 1.7 1.5 1.3 / 140%) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear -0.7 -0.5 -0.3) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear -0.7 -0.5 -0.3) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear -0.7 -0.5 -0.3 / -40%) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear -0.7 -0.5 -0.3 / -40%) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear calc(r) calc(g) calc(b))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear calc(r) calc(g) calc(b) / calc(alpha))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear none none none)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear none none none / none)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g none)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g none / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / none)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g none / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / none)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear none none none) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear none none none / none) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 none 0.3) display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color(display-p3-linear 0.7 0.5 0.3 / none) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from currentColor display-p3-linear r g b)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color(from color-mix(in display-p3-linear, color(display-p3-linear 0.7 0.5 0.3), color(display-p3-linear 0.7 0.5 0.3)) display-p3-linear r g b / alpha)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
